New Development: No             JIRA Issue: CAS-4417         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: test_flagcmd   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Added new test_plot_empty to catch the tb.getcol() seg fault.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@20719 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/python/scripts/tests/test_flagcmd.py
@@ -573,29 +573,31 @@ class test_rflag(test_base):
 
         self.assertEqual(res6['flagged'], 42728.0)
 
+
 class test_actions(test_base):
     
     def setUp(self):
         self.setUp_data4rflag()
         
     def tearDown(self):
-        os.remove('fourplot.png')
+        if os.path.exists('fourplot.png'):
+            os.remove('fourplot.png')
         
+    def test_plot_empty(self):
+        '''flagcmd: Test action=plot. Nothing to plot'''        
+        outplot = 'fourplot.png'
+        flagcmd(vis=self.vis, inpmode='table', useapplied=True, action='plot',
+                plotfile=outplot)
+        
+        self.assertFalse(os.path.exists(outplot),'Plot file should not exist')
+
     def test_plot(self):
         '''flagcmd: Test action=plot'''
-        
-        # Note: the following seg faults on active ... because
-        # tb.getcol() on an empty column has a problem. Uncomment when
-        # the problem is fixed
         outplot = 'fourplot.png'
-#        flagcmd(vis=self.vis, inpmode='table', useapplied=True, action='plot',
-#                plotfile=outplot)
-#        
-#        self.assertTrue(os.path.exists(outplot),'Plot file was not created')
-        
         flagcmd(vis=self.vis, inpmode='cmd', 
             command=["intent='CAL*POINT*' field=''","scan='5'"], 
             action='list', savepars=True)
+        
         flagcmd(vis=self.vis, inpmode='table', useapplied=True, action='plot',
                 plotfile=outplot)
         
@@ -614,7 +616,9 @@ class cleanup(test_base):
         os.system('rm -rf multiobs.ms*')
         os.system('rm -rf flagdatatest-alma.ms*')
         os.system('rm -rf ngc5921.ms*')
-        os.system('rm -rf Four_ants.ms*')
+        os.system('rm -rf Four_ants*.ms*')
+        os.system('rm -rf shadowtest*.ms*')
+        os.system('rm -rf tosr0001_scan3*.ms*')
 
 
     def test1(self):
